Move sync and clear data tests to a dedicated test class
